+++ docs/versionhistory.rst
@@ -9,6 +9,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   in ``anyio.Path``, newly added in Python 3.13
 - Changed the ``ResourceWarning`` from an unclosed memory object stream to include its
   address for easier identification
+- Fixed ``to_process.run_sync()`` failing to initialize if ``__main__.__file__`` pointed
+  to a file in a nonexistent directory
+  (`#696 <https://github.com/agronholm/anyio/issues/696>`_)
 
 **4.4.0**